Feat/support appear (#517)
* feat: support appear component

* chore: modify version

* fix: should disappear

* chore: export VisibilityChangeEvent

* feat: rax-compat use @ice/appear

* feat: need not give childref

* chore: update lock

* chore: update README and package

* chore: update README

* test: add test for appear

* chore: update lock
