+++ packages/appear/CHANGELOG.md
@@ -1,0 +1,5 @@
+# @ice/appear
+
+## v0.1.0
+
+- [feat] support VisibilityChange.
