+++ packages/appear/src/type.ts
@@ -1,0 +1,3 @@
+export function isFunction(obj: any): obj is Function {
+  return typeof obj === 'function';
+}
